#0: Fix error during post-processing of Bert tests

+++ tests/models/bert/test_bert.py
@@ -217,7 +217,7 @@ def run_bert_question_and_answering_inference(
         pytorch_out = hugging_face_reference_model(bert_input)
         tt_out = tt_bert_model(bert_input).cpu()
     tt_untilized_output = (
-        tt_out.to(ttl.tensor.Layout.ROW_MAJOR).to_torch().reshape(batch, 1, seq_len, -1)
+        tt_out.to(ttl.tensor.Layout.ROW_MAJOR).to_torch().float().reshape(batch, 1, seq_len, -1)
     )
 
     tt_start_logits = tt_untilized_output[..., :, 0].squeeze(1)
